+++ lib/Search/FolderSearchProvider.php
@@ -60,7 +60,12 @@ class FolderSearchProvider implements IProvider
     public function search(IUser $user, ISearchQuery $query): SearchResult
     {
         $list = [];
-        $term = strtolower($query->getTerm());
+        if (method_exists($query, 'getFilter')) {
+            $term = $query->getFilter('term')?->get() ?? '';
+        } else {
+            $term = $query->getTerm();
+        }
+        $term = strtolower($term);
 
         foreach ($this->service->findAllForUser($user->getUID()) as $folder) {
             if (strpos(strtolower($folder->getName()), $term) === false) {